+++ src/com/clouway/list/AdderThread.java
@@ -3,23 +3,25 @@ package com.clouway.list;
 /**
  * @author Stanislava Kaukova(sisiivanovva@gmail.com)
  */
-public class Consumer extends Thread {
+public class AdderThread extends Thread {
     private ListOfElements list;
 
-    public Consumer(ListOfElements list) {
-
+    public AdderThread(ListOfElements list) {
         this.list = list;
     }
 
     @Override
     public void run() {
-        while (true) {
+        String[] elements = {"one", "two", "three", "four", "five"};
+        for (String each : elements) {
             try {
-                String element = list.remove();
-                System.out.println("Consumer" + " " + element + "is take");
+                list.add(each);
+                System.out.println("AdderThread: " + each + " is added");
+
                 Thread.sleep(1000L);
             } catch (InterruptedException e) {
                 e.printStackTrace();
+                return;
             }
         }
     }
