+++ src/com/clouway/list/ListOfElements.java
@@ -8,31 +8,34 @@ import java.util.List;
  */
 public class ListOfElements {
     private int maxCapacity;
-    private List<String> list;
+    private List<String> list = new ArrayList<String>();
 
     public ListOfElements(int maxCapacity) {
-
         this.maxCapacity = maxCapacity;
-        list = new ArrayList<String>();
     }
 
     public synchronized void add(String element) throws InterruptedException {
         while (list.size() == maxCapacity) {
             System.out.println("The list is full. Please, wait!");
+
             wait();
         }
         list.add(element);
+
         notifyAll();
     }
 
     public synchronized String remove() throws InterruptedException {
         while (list.isEmpty()) {
             System.out.println("The list is empty. Please, wait!");
+
             wait();
         }
         String lastElement = list.get(list.size() - 1);
         list.remove(lastElement);
+
         notifyAll();
+
         return lastElement;
     }
 }
